requirements.txt
Flask
python-dotenv
langchain
pandas
chardet